rename kortex2 -> kortex
This package was ported to ros2 an some these packages were named kortex2
The 2 belongs to ros not the kortex version.

remove all kortex2 in and replace with just kortex
