+++ .github/workflows/ci-ros-lint.yml
@@ -18,8 +18,8 @@ jobs:
         distribution: rolling
         linter: ${{ matrix.linter }}
         package-name:
-          kortex2_bringup
-          kortex2_driver
+          kortex_bringup
+          kortex_driver
 
 
   ament_lint_100:
@@ -38,5 +38,5 @@ jobs:
         linter: cpplint
         arguments: "--linelength=100 --filter=-whitespace/newline"
         package-name:
-          kortex2_bringup
-          kortex2_driver
+          kortex_bringup
+          kortex_driver
