+++ .github/workflows/humble-source-build.yml
@@ -29,8 +29,8 @@ jobs:
           import-token: ${{ secrets.GITHUB_TOKEN }}
           # build all packages listed in the meta package
           package-name:
-            kortex2_bringup
-            kortex2_driver
+            kortex_bringup
+            kortex_driver
           vcs-repo-file-url: |
             https://raw.githubusercontent.com/ros2/ros2/${{ env.ROS_DISTRO }}/ros2.repos
             file://${{ github.workspace }}/ros2_kortex.${{ env.ROS_DISTRO }}.repos
