+++ kortex_bringup/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kortex2_bringup</name>
+  <name>kortex_bringup</name>
   <version>0.1.0</version>
   <description>Launch file and run-time configurations, e.g. controllers.</description>
