+++ kortex_driver/CMakeLists.txt
@@ -1,5 +1,5 @@
 cmake_minimum_required(VERSION 3.5)
-project(kortex2_driver)
+project(kortex_driver)
 
 # Default to C++14
 if(NOT CMAKE_CXX_STANDARD)
