+++ kortex_driver/package.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0"?>
 <?xml-model href="http://download.ros.org/schema/package_format3.xsd" schematypens="http://www.w3.org/2001/XMLSchema"?>
 <package format="3">
-  <name>kortex2_driver</name>
+  <name>kortex_driver</name>
   <version>0.0.0</version>
   <description>ROS2 driver package for the Kinova Robot Hardware.</description>
   <maintainer email="marq.rasmussen@picknik.ai">Marq Rasmussen</maintainer>
